rename and pass norminette except for builtin
