+++ libft/ft_calc_back_chr.c
@@ -1,34 +1,25 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   ft_strrchr.c                                       :+:      :+:    :+:   */
+/*   ft_calc_back_chr.c                                 :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: kaisobe <kaisobe@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2024/10/27 18:20:44 by tkitago           #+#    #+#             */
-/*   Updated: 2025/02/10 15:31:33 by kaisobe          ###   ########.fr       */
+/*   Created: 2025/02/10 14:45:03 by kaisobe           #+#    #+#             */
+/*   Updated: 2025/02/10 14:45:46 by kaisobe          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-char	*ft_strrchr(const char *s, int c)
+int	ft_calc_back_chr(char *str, char c)
 {
-	int	i;
+	char	*s;
 
-	i = 0;
-	if (!s)
-		return (NULL);
-	while (s[i] != '\0')
-		i++;
-	if (c == '\0')
-		return ((char *)&s[i]);
-	while (i >= 0)
+	s = ft_strrchr(str, c);
+	if (s == NULL)
 	{
-		if (s[i] == (char)c)
-			return ((char *)(s + i));
-		i--;
+		return (-1);
 	}
-	return (NULL);
+	return (s - str);
 }
-//指定された文字cが文字列sの中で最後に現れる場所を探し、その位置のポインタを返す
