--- src/executer/process_util.c
@@ -1,9 +1,0 @@
-#include "minishell.h"
-
-int	contain_backslash(char *str)
-{
-	if (ft_strchr(str, '/'))
-		return (1);
-	return (0);
-}
-
