+++ src/includes/exec.h
@@ -1,3 +1,15 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   exec.h                                             :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: kaisobe <kaisobe@student.42.fr>            +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2025/02/10 15:25:13 by kaisobe           #+#    #+#             */
+/*   Updated: 2025/02/10 15:27:33 by kaisobe          ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #ifndef EXEC_H
 # define EXEC_H
 
@@ -29,5 +41,6 @@ t_status			execute_one_builtin(t_astnode *root);
 t_status			execute_fork_commands(t_list *cmds);
 t_status			waitpids(t_list *cmds, pid_t *pids);
 t_status			check_fds(t_redirect *redirect);
+void				handle_io(t_token *redirect);
 
 #endif
